Optimization - Lifts the core metadata, MaxAge and MaxCount to the streams table. This allows this data to be read at the time of reading a stream instead of requiring an additional query. Fixes #112.
As this needs a new schema for MSSQL, this is implemented as a separate schema side-by-side with the previous version which has been marked obsolete. This allows users to use the old schema and plan a migration independent of a package update. The old version will be removed in a future version.